backend/requirements.txt
Django==5.2.5
djangorestframework==3.16.1
django-cors-headers==4.7.0
channels==4.3.1
channels-redis==4.3.0
redis==6.3.0
uvicorn==0.35.0
gunicorn==23.0.0
ffmpeg-python==0.2.0
opencv-python-headless==4.12.0.88
pillow==11.3.0
numpy==2.2.6
psycopg2-binary==2.9.10
dj-database-url==3.0.1
python-dotenv==1.1.1
websockets==15.0.1